backend/pyproject.toml
[project]
name = "papermap"
version = "0.1.0"
description = "Semantic paper explorer - visualize arxiv research landscape"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
]
